+++ CHANGELOG.rst
@@ -20,6 +20,7 @@ Fixed
 ^^^^^
 - ``data``: Copernicus Open Access Hub Url changed #142
 - ``data``: fixed download error from EarthExplorer #140
+- ``data``: Removed pylandsat dependency and added methods for downloading landsat products from GCS in data module #106
 
 [1.2.0]] (2021-04-07)
 ---------------------